requirements.txt
streamlit>=1.35.0
anthropic>=0.40.0
snowflake-connector-python>=3.10.0
pandas>=2.0.0
plotly>=5.0.0
